src.ctf.fs: Implement queries with new `MetadataStreamDecoder' and `Ctf1MetadataStreamParser`
Rewrite the `babeltrace.support-info` query using the new
`Ctf1MetadataStreamParser` class.

Rewrite the `metadata-info` query using the new `MetadataStreamDecoder`
class.

FIXME: when TMPDIR gets long, the error messages get wrapped, added
workaround to remove \r\n.  The problem appears with this commit,
because before, we would get a log line (at the error level) with the
same message, so grep would find something even if the message in the
error stack is wrapped.  The new code throws an error, but does not log
an error.  I think that's what we want, because in some cases we want to
swallow the error and we don't want a errors to be logged.  Before
merging, find a better solution for this.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Change-Id: I80530c6eef7ca62c4271a568e527e48e8ec2602a
Reviewed-on: https://review.lttng.org/c/babeltrace/+/7946
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>