doc/api/libbabeltrace2/Makefile.am: depend on all source files
This patch adds dependencies to the `$(DX_DOCDIR)/$(PACKAGE).tag`
target, which is a tag file which the Doxygen process generates when
done.

The added dependencies are `Doxyfile` (the configuration),
`DoxygenLayout.xml` (layout), `style.css` (custom style), everything
under `dox` (general documentation), and all the public headers.

The effect is that changing any of those files make the `doxygen-doc`
target regenerate the whole API documentation project.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: Ic8f686ab08f2047290f3d4f5b20daa2e12b85447