doc: libbabeltrace2: Doxyfile: set `GENERATE_TREEVIEW` to `YES`
The main reason I didn't enable the tree view is because I had an
annoying Doxygen bug in Chromium, which I reported [1].

This bug is now fixed starting from Doxygen 1.8.17, therefore we can
enable the tree view now, which in my opinion increases the usability of
the API documentation system.

That being said, there's still one annoying behaviour with the tree view
enabled: when you refresh the page without a current anchor (`#`), the
scroll position goes back to the top (in both Chromium and Firefox).
However, this is mostly annoying to me who'll often refresh the page to
render the updated documentation when actively writing it. For this use
case, I can disable the tree view locally when writing
the documentation.

[1]: https://github.com/doxygen/doxygen/issues/7412

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: Iba9984093ce79d70dc5612c7c02ef6a373c26c97