bt2: make some types `enum.Enum`
Make a few classes inherit from `enum.Enum`.

The return type annotation for method `_FieldPathConst.root_scope()`,
for instance,  says that it returns an `FieldPathScope`, but it's a lie:
it's `int`.  The `FieldPathScope` class only serves to hold a few `int`
static fields right now.  However, we can make `FieldPathScope` an enum
(`enum.Enum`), and that becomes true.

Users are expected not to rely on the integer value of the enumerators.
They are expected to use the enumerator, for instance
`bt2.IntegerDisplayBase.DECIMAL`.  However, there is actually one
instance where we rely on the integer value in the tests, in
`tests/data/plugins/src.ctf.fs/field/bt_plugin_test_text.py`:

    elif isinstance(field.cls, bt2._IntegerFieldClassConst):
        # Honor the preferred display base
        base = field.cls.preferred_display_base
        print(indent_str + intro_str, end="")

        if base == 10:
            print(int(field))
        elif base == 16:
            print(hex(field))
        elif base == 8:
            print(oct(field))
        elif base == 2:
            print(bin(field))

If we make the `IntegerDisplayBase` an `enum.Enum`, then all these
comparisons will always be false (`IntegerDisplayBase.DECIMAL` will not
be equal to the int value 10).  We could say "this is user error", fix
the test and go on, but it's not impossible for user code out there to
do the same thing.  Make `IntegerDisplayBase` be an `enum.IntEnum` to
make these comparisons work.

It's also equally likely for that someone is passing an int value when
creating an integer field class:

    tc.create_unsigned_integer_field_class(preferred_display_base=16)

So, in `_IntegerFieldClass._set_preferred_display_base`, allow for
`base` to be an integer.  Trying to construct an `IntegerDisplayBase`
from an invalid int value raises `ValueError`, which preserves the
behavior we had before, where we did a manual check.

I think the `IntegerDisplayBase` case is a bit special, where the
enumerator values match what they mean.  For other enumerations, I don't
think it's that much of a problem.  Even though it's possible in theory,
I can't imagine someone doing

    if log_level == 2:

instead of the more appropriate

    if log_level == bt2.LoggingLevel.DEBUG

For that reason, I did not use `enum.IntEnum` everywhere, but we can
always do it if someone complains.

Except for `IntegerDisplayBase`, remove some tests that were checking
that passing a wrong int value would raise a `ValueError`.  We now check
that the passed values have the proper enum type, so passing anything
else than the proper enum type will raise a `TypeError`.

Change-Id: I8369112fe37c2d12371deb3b20ad5c1fa1b97823
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13566
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>