Fix: cpp-common/bt2, clock-correlation-validator: add two missing `!`
They are both somewhat obvious, causing the clock classes and origins
not to get compared correctly.  They somehow cancelled each other
out most of the time, which is why they passed under the radar.

I took the opportunity to add some success tests to
`test-clock-compatibility.cpp`.  The structural changes of this test
are:

 - move the functionality of sending messages with two different clock
   classes to a base class `TestCase`
 - `TestCase::_runOne()` passes to the child class'
   `_handleResult()` method whether running the graph threw an
   exception, and the current thread's error, if any
 - the two child classes, in their respective `_handleResult()` methods,
   verify the appropriate assertions

Change-Id: I691c8b50e314bc1a28b106ea719ebf30686940fc
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13553
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>