bt2: define overloads of `Graph.add_component()`
Define overloads for source, filter and sink.  This lets analysis tools
know that if you pass a source component class, the method returns a
source component (and likewise for the other types).

The `noqa: F811` comments are necessary to avoid the following flake8
warnings:

    $ flake8
    ./src/bindings/python/bt2/bt2/graph.py:74:5: F811 redefinition of unused 'add_component' from line 60
    ./src/bindings/python/bt2/bt2/graph.py:88:5: F811 redefinition of unused 'add_component' from line 74
    ./src/bindings/python/bt2/bt2/graph.py:101:5: F811 redefinition of unused 'add_component' from line 88

I believe this is due to our typing module hack resulting in flake8 not
recognizing `typing.overload`: I don't get these warnings when using
`import typing` at the top of the file, rather than our hack.  So in the
distant future, when we can use the built-in typing module directly, we
can remove these comments.

Change-Id: Idd8bfe02053de7f2bb0dd71153f83de250e89843
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13561
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>