bt2: add typing annotations for values, params
Define `_ConvertibleTo*Value` types, representing the input types that
can be converted to the different kind of values, through
`bt2.create_value()`.

Define `_ComponentParams` as the types a user can pass to a component.
Annotate a few `params` parameters with it.

Define `_QueryParams` as the types a user can pass to a query.  Annotate
the `params` parameter of `QueryExecutor.__init__()` with it.

Change-Id: I12251c2269382a51df9b9f4650729611c43a5bdb
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13562
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>
Tested-by: jenkins <jenkins@lttng.org>