tests: add "all fields" source
Move the plugin `bt_plugin_test_debug_info.py`, currently specific to
the `flt.lttng-utils.debug-info` tests, to
`tests/utils/python/bt_plugin_trace_ir_test.py`, to make it available to
other tests.  This plugin defines one source that sends an event message
with a variety of fields.  This source has nothing specific to
`flt.lttng-utils.debug-info`, and I think it could be useful for other
tests.

Rename the plugin to `trace-ir-test`.  The intent is that other sources
could be added to cover other corners of the trace IR.

Do minor adjustments to the code to make it clean, typing-wise.

Remove the clock class, as it's not needed to test fields.

Change-Id: Ieb91ce5e7a8e70676f2fd4b7937c4347d5970620
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>