include/babeltrace2: add MIP version support precondition when adding component to graph
While adjusting component classes to support MIP 1, we determined that
since nothing prevents adding a component to a graph with an effective
MIP version the component doesn't support, it would be advisable to put
a MIP version check in the component's initialize method, returning an
error if it doesn't supported the graph's effective MIP version.  Since,
to be consistent and rigorous, this check would be needed in all
component classes ever written, the question is: why doesn't the library
check that instead?

As of now, the "get supported MIP versions" method of component classes
is optional.  It is only required to implement it if you want your
component class to be able to play well with
`bt_get_greatest_operative_mip_version()`, which is used for instance by
the Babeltrace 2 CLI to automatically configure a graph.  If you don't
need your component class to participate in this (for instance, you
always configure your graph involving this component class explicitly),
it is valid today for your component class to effectively support MIP 1,
but leave the "get supported MIP versions" unimplemented (which means
it implicitly returns MIP 0 as the sole supported version).

I propose to make it required for component classes to implement the
"get supported MIP versions" method (unless they only support MIP 0, in
which case the implicit behavior is fine) to indicate which versions
they effectively support, such that the library becomes able to verify
that components added to a graph do indeed support the graph's effective
MIP version.

I believe that this change is safe to do right now, when all existing
component classes in the universe can only support MIP 0.  All graphs
created today use MIP 0 (there's a precondition for that).  All
component classes in existence probably* have a "get supported MIP
versions" method that returns a range set that includes MIP 0.  So if we
start checking today that the graph's effective MIP version is supported
by the component classes, the code written with Babeltrace 2.0 in mind
should always pass that check.

* I say probably because it's technically possible for someone today to
  implement the method to return `[7]`, while effectively supporting MIP
  0.  But I would consider this trolling and not really care about that.

I believe that this change would be unsafe to do later.  Once Babeltrace
2.1 is released and MIPs 0 and 1 are in the wild, one could implement a
component class that effectively supports only MIP 1, but leave "get
supported MIP versions" unimplemented (because they do not care for it).
If we were to introduce the precondition with, say Babeltrace 2.2 and
MIP 2, then that component class that worked with Babeltrace 2.1 would
start failing the check with Babeltrace 2.2.  So, it is now or never 🎵.

The important thing is to get the precondition in for Babeltrace 2.1.
We can implement the actual check for Babeltrace 2.1 or later, it
doesn't matter.

So, this patch adds preconditions to the
`bt_graph_add_*_component_with_initialize_method_data()` to verify
this.  We never say explicitly that it is mandatory to implement the
"get supported MIP versions" method if you want your component class to
supported something else that just 0, but this precondition implicitly
makes it mandatory.  However, it would probably be good to modify the
"component class development" chapter of the doc to clarify that.

Philippe updated the documentation.

Change-Id: I9fb5171a368bb9b5f4e1505889f22857390ef7d3
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13652