src.ctf.fs: consider namespace, name and uid when merging traces
To get a truly unique identifier for a trace, we must use the name and
(if present) the namespace.

Update the `babeltrace.support-info` query handling of `src.ctf.fs` to
only emit a `group` string if both name and uid are present.

Update the check in `ctf_fs_component_create_ctf_fs_trace()` that
verifies that all traces passed as inputs share the same identity to use
the same logic.

Add an implicit constructor to `bt2c::CStringView` to allow constructing
from a `bt2s::optional<std::string>`, which is handy here:

    const bt2::IdentityView firstTraceId {firstTrace->cls()->ns(), firstTrace->cls()->name(),
                                          firstTrace->cls()->uid()};

Change-Id: I7f945a7719f1c7cbfe91f5b3295b6197c303f51f
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>