Fix: src.ctf.*: detect empty event record and throw
A 0-bit event record is invalid because it hinders decoding progress.

Without this patch, for this trace:

    <RS>
    {
      "type": "preamble",
      "version": 2
    }
    <RS>
    {
      "type": "data-stream-class"
    }
    <RS>
    {
      "type": "event-record-class"
    }

and any data stream, you get an infinite stream of:

    <unknown>:
    <unknown>:
    <unknown>:
    <unknown>:
    <unknown>:
    <unknown>:
    <unknown>:
    <unknown>:
    ...

With this patch, you get:

    ERROR:    [Babeltrace CLI]
      Graph failed to complete successfully
    CAUSED BY [libbabeltrace2]
      Component's "consume" method failed
    CAUSED BY [libbabeltrace2]
      Component input port message iterator's "next" method failed
    CAUSED BY [libbabeltrace2]
      Component input port message iterator's "next" method failed
    CAUSED BY [auto-disc-source-ctf-fs ...]
      Failed to create the next message
    CAUSED BY [auto-disc-source-ctf-fs ...]
      At 0 bits: Invalid empty event record.

In `ctf::src::ItemSeqIter`, I'm just keeping a snapshot of the head
position within the current packet at the beginning of any
event record E and comparing it to the head position at the end of E.
If both are equal, then it's an empty, invalid event record.

This is sometimes detectable at the metadata level, but sometimes not,
for example if the contents of an event record is an optional field of
which the selector is located within the packet context. Therefore I'm
choosing the safest approach now.

Adding a `src.ctf.fs` failure test to check this.

Fixes: https://support.efficios.com/issues/1439
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I9bb701369fe31842eeb40a19f0c4bf664133465c
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13798
Tested-by: jenkins <jenkins@lttng.org>