lib: remove internal ns/name/uid strings for Unix epoch clock origin
The Unix epoch origin is currently represented by the library with the
following constants:

    static const char *const CC_ORIGIN_UNIX_EPOCH_NS = "babeltrace.org,2020";
    static const char *const CC_ORIGIN_UNIX_EPOCH_NAME = "unix-epoch";
    static const char *const CC_ORIGIN_UNIX_EPOCH_UID = "";

These strings can be read by the user with
`bt_clock_class_get_origin_{namespace,name,uid}()` but it is made clear
in the documentation that these values should not be relied upon by the
user.

This brought confusion, for instance, it wasn't clear if the user should
be allowed to call `bt_clock_class_set_origin()` with those values, and
if that should result in a clock class with the Unix epoch origin.

To avoid confusion, remove those strings entirely, and only allow
`bt_clock_class_get_origin_{namespace,name,uid}()` to be called if the
origin has been set with `bt_clock_class_set_origin()`, meaning the
origin is not the Unix epoch and is not unknown (i.e. is "custom").

In the library, represent the clock origin kind with an enum, which
makes things quite clear and straightforward.

Update some other bits of code following the new restrictions imposed on
`bt_clock_class_get_origin_{namespace,name,uid}()`.

Documentation updated by P. Proulx.

Change-Id: Ic042cb5b9b2592baf30b4647ece5249b62f70593
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>