src.ctf.*: set CTF IR trace name of CTF 1.8 traces to empty string
This is to ensure there's at least a name so that merging traces by
(namespace, name, UID) tuple is possible (name and UID are required
for this).

It's okay to have such an empty name here since for a CTF 1.8 trace,
the UID is a UUID which already is globally unique.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I0ec0b636450fcd03e0c29424257e5129fa0e1d74
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13818
Tested-by: jenkins <jenkins@lttng.org>