src.ctf.fs: use trace namespace and name to make port name
This patch changes the ctf_fs_make_port_name() function so that it not
only considers the trace UID to make the port name, but also the trace
namespace (if available) and name.

Because the trace UID is only unique within the context of the trace
namespace and name, `src.ctf.fs` requires both the name and UID to
coexist to consider it. This is the same strategy as the one to merge
physical traces into a logical trace.

When both the trace name and UID exist, instead of starting only with
the UID, the new port name format starts with

    {namespace: `NS`, name: `NAME`, uid: `UID`}

or

    {name: `NAME`, uid: `UID`}

Here's a real port name example:

    {namespace: `lttng.org,2009`, name: ``, uid: `6f180b0c-b242-c148-ab44-6cbf960a58b2`} | 1 | 3

`tests/plugins/src.ctf.fs/query/test_query_trace_info.py` tests port
names for both CTF 1.8 and CTF 2. Therefore, I'm updating the two
relevant CTF 2 traces to add an empty trace name so that the port name
remains the same for both CTF versions. Also changing what's expected.

Updating babeltrace2-source.ctf.fs(7) to document the new port
name format.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I39ea934683ce6458a4474fc152b5bfe141b3cf84