Fix: src.ctf.fs: ensure index is correctly sorted when clock offset changes
Issue observed
--------------

Babeltrace fails to decode a set of trace chunks with the following
error:

  01-14 13:49:55.957 98397 98397 F LIB/ASSERT-COND bt_lib_assert_cond_failed@lib/assert-cond.c:63 Babeltrace 2 library postcondition not satisfied.
  01-14 13:49:55.957 98397 98397 F LIB/ASSERT-COND bt_lib_assert_cond_failed@lib/assert-cond.c:65 ------------------------------------------------------------------------
  01-14 13:49:55.957 98397 98397 F LIB/ASSERT-COND bt_lib_assert_cond_failed@lib/assert-cond.c:66 Condition ID: `post:message-iterator-class-next-method:message-clock-snapshots-are-monotonic`.
  01-14 13:49:55.957 98397 98397 F LIB/ASSERT-COND bt_lib_assert_cond_failed@lib/assert-cond.c:68 Function: bt_message_iterator_class_next_method().
  01-14 13:49:55.957 98397 98397 F LIB/ASSERT-COND bt_lib_assert_cond_failed@lib/assert-cond.c:69 ------------------------------------------------------------------------
  01-14 13:49:55.957 98397 98397 F LIB/ASSERT-COND bt_lib_assert_cond_failed@lib/assert-cond.c:70 Error is:
  01-14 13:49:55.957 98397 98397 F LIB/ASSERT-COND bt_lib_assert_cond_failed@lib/assert-cond.c:72 Clock snapshots are not monotonic
  01-14 13:49:55.957 98397 98397 F LIB/ASSERT-COND bt_lib_assert_cond_failed@lib/assert-cond.c:75 Aborting...

Cause
-----

Inspecting the metadata of the various trace chunks, it becomes apparent
that the clock class declared for the trace changes offset over time.
This happens when LTTng re-samples the clock offset between two session
rotations.

Each index entry contains `timestamp_begin_ns` and `timestamp_end_ns`
fields.  These fields are computed using the chunk's clock class offset,
and are used to keep the entries sorted chronologically.  When the clock
class offset changes between chunks, this can lead to out of order index
entries.  The actual decoding of the trace is done using the clock class
(and its offset) of one "winner" chunk.  Eventually, when decoding
reaches the out of place index entry, we see the error above.

Here's an example with made up numbers.  Consider a stream with two
chunks, one packet each.  A clock resampling occured between the two
chunks were written, changing the clock offset from 1000 to 500 ns.  The
built index would look like this, with the packets out of order, sorted
by timestamp begin (ns):

 - Packet 2, in chunk 2
   - Clock class offset from origin (ns): 500
   - timestamp begin (cycles): 30
   - timestamp end (cycles): 40
   - timestamp begin (ns): 530
   - timestamp end (ns): 540

 - Packet 1, in chunk 1
   - Clock class offset from origin (ns): 1000
   - timestamp begin (cycles): 10
   - timestamp end (cycles): 20
   - timestamp begin (ns): 1010
   - timestamp end (ns): 1020

When doing the actual decoding of the trace, let's say that the clock
class from chunk 2 is used.  The message iterator first emits a packet
with bounds (530 ns, 540 ns), corresponding to packet 2.  It then emits
a packet with bounds (510 ns, 520 ns), corresponding to packet 1, since
it decoded that packet using chunk 2's clock class.  This results
non-monotonic clock snapshots, and the assertion error above.

Solution
--------

Sort index entries using the raw timestamps, not the timestamp converted
to nanoseconds.  This results in the correct order in the index.

Add a test that reproduces more or less the example described above.

Known drawbacks
---------------

The choice of offset-to-epoch that is used is somewhat arbitrary: it is
the one from the "winner" metadata, which simply means it is the one
that contains the most event classes (IOW, "the most complete").

From the reader's perspective, it is hard to determine the "best" offset
to use. In practice, we mostly want to use one that doesn't place the
trace in 1970. There is no logic that ensures that right now, so I
assume it is possible the elected metadata could be the one generated
during the capture of the first trace chunk before a major NTP
correction could take place assuming it contains all event classes that
will ever be recorded for that session.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I6892e71c5b1e760efd19d9835b1a2ea54dc79b00
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13850
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>