Fix: ctf: handle non-map value params in queries
Unlike for component parameters, the top-level values used for query
parameters do not have to be maps.  The `src.ctf.fs` and
`src.ctf.lttng-live` component classes do not handle receive anything
else than a map correctly.

We don't see the issue when invoked from the command line, because the
CLI only knows how to pass maps as the top-level value of parameters.
It is however possible to reproduce using the C or Python API.

All queries use the param validation utility, so add a check in that
utility.  Given that query params don't _have_ to be a map, at the
top-level, I suppose the param validation utility could be improved to
support having other value types as the top-level.  It just isn't needed
right now.

Add some tests to cover passing null values to all these queries.

Change-Id: Id3ff031f93948bbb97c7c722ccb1a80d023fa197
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13857
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>
Tested-by: jenkins <jenkins@lttng.org>