bt2: return active flags instead of flag labels
To be similar to `_EnumerationFieldClassConst.mappings_for_value()`,
rename
`_BitArrayFieldClassConst.active_flag_labels_for_value_as_integer()` to
`_BitArrayFieldClassConst.active_flags_for_value_as_integer()` and make
it return a list of flags, instead of a list of labels.

`_EnumerationFieldConst.labels()` returns a list of labels, so leave
`_BitArrayFieldConst.active_flag_labels()` the way it is.

Change-Id: Icaf549f44e5794100004c75d7902b9f2fc7e3d0a
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13873
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>
Tested-by: jenkins <jenkins@lttng.org>