cpp-common/bt2: add null checks in shared object get/put ref functions
Functions `RefFuncsT::get()` and `RefFuncsT::put()` are called
unconditionally by class `SharedObject`, because they are expected to be
able to handle `nullptr` / `NULL` values.  However, there appears to be
a significant cost to do these function calls when unnecessary (when the
internal pointer is `nullptr`).  To avoid that cost, add some checks in
`SharedObject`.  These checks are likely to be inlined in the callers,
so will avoid branching out of the calling functions.

 - Add some NULL checks in `_getRef()` and `_putRef()`.

 - Add `_getRefNoNullCheck()`, for use in `createWithRef()`, where we
know the `SharedObject` isn't empty.

 - Remove the check in `reset()`, since that check is now redundant.

On my test trace, times go from:

    Run 1: Real: 84.84 User: 84.44 Sys: 0.39
    Run 2: Real: 85.13 User: 84.74 Sys: 0.37

to

    Run 1: Real: 79.57 User: 79.18 Sys: 0.38
    Run 2: Real: 79.69 User: 79.29 Sys: 0.38

Change-Id: I156593c167b2fe0f01da9d8ffe1decd7aa690cbf
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13894
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>
Tested-by: jenkins <jenkins@lttng.org>