src.ctf.fs: remove `ctf_fs_ds_file_info::beginNs()`
The `_mBeginNs` field of `ctf_fs_ds_file_info` is never set.  Yet, it is
based on that field that `ctf_fs_ds_file_info` objects are kept sorted,
in `ctf_fs_ds_file_group::insert_ds_file_info_sorted()`.

The field was set prior to commit 9328f86be150 ("src.ctf.fs: use the new
metadata stream parser and message iterator").

But it turns out that it's harmless, because the list of
`ctf_fs_ds_file_info` objects doesn't need to be sorted.  The important
thing that the index (`ctf_fs_ds_file_group::index`) is sorted.
`ctf::fs::src::Medium` uses that index as a "playlist" to know which
packets to read in what order.  Each index entry contains the path to
the data stream file to read from.

Given that the order of the `ctf_fs_ds_file_info` objects is not
important, rename the `insert_ds_file_info_sorted()` method to
`add_ds_file_info()`, and make it emplace an element at the end of the
vector.

Remove the `ctf_fs_ds_file_info::_mBeginNs` field and getter.

Change-Id: I287f9726076f0624ba62cf3acaf641b413bcac71
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13910
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>
Tested-by: jenkins <jenkins@lttng.org>