src.ctf.fs: remove `ctf_fs_ds_index_entry::timestamp_{begin,end}`
Following the previous patch, the only remaining uses of the
`timestamp_begin` and `timestamp_end` is to fill the `range-ns` map in
the `babeltrace.trace-infos` query.  And even there, we use only the
begin timestamp of the first entry and the end timestamp of the last
entry.  We therefore do a lot of work that is just wasted.

Remove those fields, all the code to compute them, and compute the
timestamp in nanoseconds from origin on demand when handling the trace
infos query, in `populate_stream_info()`.

Change-Id: I29cb1baaa3ca0652e1674ab381840dbbf1687d97
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13913
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>
Tested-by: jenkins <jenkins@lttng.org>