Fix: Python: missing __contains__ method for _PluginComponentClasses
The python binding tests fail on Python 3.4 and 3.5 when attempting to look up a
component class with an invalid name
(`test_create_from_name_wrong_component_class_name_type`) as as unexpected
`KeyError` is raised instead of the expected `TypeError`.

In these older Python versions, `typing.Mapping` does not provide a
`__contains__` method, which is used by the `in` operator. Instead, the
interpreter falls back to using the iterator to sequentially compare the
requested key against all values.

In more recent Python versions, the `__contains__` method is provided by
`typing.Mapping` and the test relies on this method to call `__getitem__` and
fail with a `TypeError` since the name of a component class must be a string
(`_check_str` fails).

This commit adds a `__contains__` method to `_PluginComponentClasses` that uses
`__getitem__` to check for the presence of a key and raises the `TypeError` that
test excepts.

Change-Id: I97562903154831e7e5dd0e1ab581eb5d9d9510d1
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/14610