tests: add bt_test_cli
Add the `bt_test_cli` test util function, whose purpose is to reduce the
boilerplate in tests.  `bt_test_cli` runs `bt_cli`, checks some aspects
of the result, and records some TAP test results (hence the `test` in
the name).  See its doc for more details.

I changed a bunch of tests to use it, just enough to validate that the
approach was flexible enough and can apply in a variety of situations.
There are certainly more tests that could be changed to use it, and we
can add more options later to cater to more use cases.

Change-Id: I54d3b4ae96d33b0a5c7eae374b30d7ef5880fe09
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/14629
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>
Tested-by: jenkins <jenkins@lttng.org>