configure: add --disable-git-version
When you:

 - build on a given commit
 - do some local changes
 - rebuild

... the generated src/common/version.i file changes to include
"-dirty".  This causes some .o files (that include it) to be rebuilt,
and many artifacts to be rebuilt.  This is slightly annoying when
working on Babeltrace itself.

To avoid that, add the `--disable-git-version` configure flag, which
ultimately instructs gen-version-i.sh to not get the version from
git.

Change-Id: I1e7b0552472ecb887558d481f0924fc7b40bbc0e
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/14663
Reviewed-by: Michael Jeanson <mjeanson@efficios.com>
Tested-by: jenkins <jenkins@lttng.org>
CI-Build: Michael Jeanson <mjeanson@efficios.com>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>