fix: m4: ensure newline after the no-op command
The no-op builtin shell command ':' is prepended to the second argument
of many AS_IF calls to ensure there is at least one command in the
'else' branch of the condition. A 'dnl' m4 macro was added after the ':'
to remove the newline from the output of the macro which results in the
first line of the marcro argument being passed as an argument to ':'
which silently ignored.

Change-Id: If1ebc60ecf0fb9c87034e39816a41baea808fb50
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/14727
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Simon Marchi <simon.marchi@efficios.com>